services/gateway/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
PyJWT[crypto]>=2.8.0
cryptography>=42.0.0
httpx>=0.27.0
redis[hiredis]>=5.0.0
aiosqlite>=0.20.0
anthropic>=0.28.0
pyyaml>=6.0.1
python-multipart>=0.0.9
prometheus-client>=0.20.0
jsonschema>=4.22.0
mmh3>=4.1.0
